+++ providers/edge3/src/airflow/providers/edge3/cli/worker.py
@@ -51,17 +51,12 @@ from airflow.providers.edge3.cli.signalling import (
     write_pid_to_pidfile,
 )
 from airflow.providers.edge3.models.edge_worker import EdgeWorkerState, EdgeWorkerVersionException
-from airflow.providers.edge3.version_compat import AIRFLOW_V_3_0_PLUS, AIRFLOW_V_3_1_PLUS
+from airflow.providers.edge3.version_compat import AIRFLOW_V_3_0_PLUS, timezone
 from airflow.utils.net import getfqdn
 from airflow.utils.state import TaskInstanceState
 
 if TYPE_CHECKING:
     from airflow.providers.edge3.worker_api.datamodels import EdgeJobFetched
-
-if AIRFLOW_V_3_1_PLUS:
-    from airflow.sdk import timezone
-else:
-    from airflow.utils import timezone  # type: ignore[no-redef,attr-defined]
 
 logger = logging.getLogger(__name__)
