+++ providers/edge3/tests/unit/edge3/cli/test_worker.py
@@ -48,12 +48,7 @@ from airflow.providers.edge3.worker_api.datamodels import (
 from airflow.utils.state import TaskInstanceState
 
 from tests_common.test_utils.config import conf_vars
-from tests_common.test_utils.version_compat import AIRFLOW_V_3_0_PLUS, AIRFLOW_V_3_1_PLUS
-
-if AIRFLOW_V_3_1_PLUS:
-    from airflow.sdk import timezone
-else:
-    from airflow.utils import timezone  # type: ignore[no-redef,attr-defined]
+from tests_common.test_utils.version_compat import AIRFLOW_V_3_0_PLUS, timezone
 
 pytest.importorskip("pydantic", minversion="2.0.0")
